* master: (21 commits)
  version info: update
  WinSCPFileSystem.cpp: Bugfix.
  TWinSCPFileSystem::MultipleEdit: Add comments.
  TWinSCPFileSystem::MultipleEdit: Refactoring.
  Refactoring: Use ClearStruct
  TWinSCPFileSystem::MultipleEdit: Refactoring.
  WinSCPFileSystem.cpp: Cleanup.
  ne_socket.c: Try to optimize performance.
  Queue.cpp: NB_DISABLE_COPY(TTerminalItem)
  version info: update.
  Bugfix: Plugin displays summary size of files incorrectly.
  FarPlugin.cpp: Cleanup.
  version info: update.
  Bugfix: NetBox commands --> Synchronize fails.
  Fix minor warnings.
  WebDAVFileSystem.cpp: Correct paths to neon includes.
  CMakeLists.txt: Remove ${LIBS_ROOT}/neon/src from global include directories.
  Fix x64 warnings.
  Fix x64 warnings.
  Fix x64 warnings.
  ...

Conflicts:
	src/NetBox/FarDialog.cpp
	src/NetBox/FarPlugin.cpp
	src/NetBox/WinSCPFileSystem.cpp

+++ src/core/ScpFileSystem.cpp
@@ -2684,11 +2684,11 @@ void TSCPFileSystem::SCPSink(const UnicodeString & FileName,
           {
             FileData.LocalFileAttrs = faArchive;
           }
-          uintptr_t NewAttrs = CopyParam->LocalFileAttrs(FileData.RemoteRights);
+          DWORD NewAttrs = CopyParam->LocalFileAttrs(FileData.RemoteRights);
           if ((NewAttrs & FileData.LocalFileAttrs) != NewAttrs)
           {
             FILE_OPERATION_LOOP (FMTLOAD(CANT_SET_ATTRS, DestFileName.c_str()),
-              THROWOSIFFALSE(FTerminal->SetLocalFileAttributes(DestFileName, (DWORD)(FileData.LocalFileAttrs | NewAttrs)) == 0);
+              THROWOSIFFALSE(FTerminal->SetLocalFileAttributes(DestFileName, FileData.LocalFileAttrs | NewAttrs) == 0);
             );
           }
         }
